requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
langchain==0.2.14
langchain-community==0.2.12
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank_bm25==0.2.2
scikit-learn==1.5.1
pydantic==2.8.2
python-dotenv==1.0.1
rich==13.7.1
openai==1.41.0
requests==2.32.3
